add pago bootcoin

+++ src/main/java/pe/com/bank/bootcoin/document/BootcoinDocument.java
@@ -18,5 +18,5 @@ public class BootcoinDocument {
     private String documentType;
     private int phoneNumber;
     private String email;
-    private int amount;
+    private Double amount;
 }
